\name{MVMR.CUE-package}
\alias{MVMR.CUE-package}
\docType{package}
\title{
Multivariable Mendelian Randomization Accounting for Correlated Horizontal Pleiotropy
}
\description{
Multivariable Mendelian randomization accounting for complex correlated horizontal
pleiotropy while elucidating shared genetic etiology.
}
\author{
Qing Cheng \email{qingcheng0218@gmail.com}
}
\keyword{package}
