\name{MVMR.CUE-package}
\alias{MVMR.CUE-package}
\alias{MVMR.CUE}
\docType{package}
\title{
\packageTitle{MVMR.CUE}
}
\description{
\packageDescription{MVMR.CUE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MVMR.CUE}
\packageIndices{MVMR.CUE}
~~ An overview of how to use the package, ~~
~~ including the most important functions ~~
}
\author{
\packageAuthor{MVMR.CUE}

Maintainer: \packageMaintainer{MVMR.CUE}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one ~~
~~ per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
